Fix 2904039 https://sourceforge.net/tracker/?func=detail&aid=2904039&group_id=216267&atid=1036998
Added Create Network to top-level window
Adding non-Network nodes to top-level window is now disallowed.

There's still the issue that if you create a network on the top-level window, you may accidentally drag it into a Network with window open. I will probably need to add the ability to move a node to it's parent (either by drag or by action).