Fixes (mostly) issue #218 - Added additional None checks to view.run function.
Original issue was (mostly) caused when the interactive mode window
closed before the simulation threads have had a chance to close. Fix
adds None checks to view.run function to check if interactive mode is
close (and thus = None) so that it does not throw an exception, and does
not cause the threads to wait for a kill which never comes.