Fixes issue #224 - Removed object association implementation in the spa/view.py code.
To solve the problem of needing to associate the bg / thal object with
the rules that go with each object, the original spa/view.py code kept a
dictionary of these associations. However, if the objects are constantly
regenerated, these associates are kept around, causing a memory leak.

The solution is not to store these associates at all, and to instead put
the object & rule information in the objects documentation
(obj.documentation). This way, no object are stored, and thus no memory
leak can occur.

This implementation of storing necessary data in the object
documentation should be used in the future, however, it should be noted
that if the object documentation is used for something else (e.g. script
generation), that this approach needs to synchronize these efforts.