Added save & restore from backup layout file to java side as well.
One thing to note: Currently, if Nengo fails to load a layout file (due to corruption of the layout file format), it will attempt to load from the backup file. If it is successful, Nengo does not overwrite the original (corrupt) layout file. The user has to physically press the "save" button to save this layout. Should this be automatic in the future? (Similar situation for the python side)